feat: Added NewHabit screen to replace create habit modal
Moving the create habit form to a new screen provides a better user experience, avoiding modal overlap with keyboard and providing scope to add extra form fields in future without being constrained by modal size.
CreateHabit component deleted as no longer necesscary, replaced with a simple button to navigate to new screen in log page.
Tests and snapshots updated to reflect changes.
