+++ src/screens/LogPage.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { ScrollView } from 'react-native';
+import { Button } from 'react-native-paper';
 //components
 import LogButton from '../components/LogButton';
-import CreateHabit from '../components/CreateHabit';
 //REDUX
 import { connect } from 'react-redux';
 
-function LogPage({ habits }) {
+function LogPage({ habits, navigation }) {
     const habitButtons = habits ? Object.keys(habits) : [];
 
     return (
         <ScrollView>
-            <CreateHabit />
+            <Button onPress={() => navigation.navigate('New')}>Create new habit</Button>
             {habitButtons.map((item) => (<LogButton id={item} key={item} />))}
         </ScrollView>
     );
